refactor BayesPairing + routing tweaks
- BayesPairing currently operates through absolute paths in-terms of
imports and file references. This means it was designed to be
run within the source folder which is incompatible with our
architecture. This commit refactors BayesPairing in-order to take
advantage of relative paths. This includes not only imports,
but instances where files are loaded.
- This also includes some tweaks to top-level routing to set-up
a basic example of a callable endpoint. This is functional.